Move and rename p2p.scala to Connect.scala (comm=>node) + extract ErrorHandler
