+++ node/src/main/scala/coop/rchain/node/connect/Connect.scala
@@ -1,4 +1,4 @@
-package coop.rchain.p2p
+package coop.rchain.node.connect
 
 import coop.rchain.p2p.effects._
 
@@ -18,10 +18,9 @@ import coop.rchain.catscontrib._, Catscontrib._, ski._
 import com.google.protobuf.ByteString
 import coop.rchain.comm.transport._, CommunicationResponse._, CommMessages._
 import coop.rchain.shared._
+import coop.rchain.comm.CommError.ErrorHandler
 
-object Network {
-
-  type ErrorHandler[F[_]] = ApplicativeError_[F, CommError]
+object Connect {
 
   val defaultTimeout: Duration = Duration(500, MILLISECONDS)
