+++ node/src/main/scala/coop/rchain/node/node.scala
@@ -29,6 +29,7 @@ import coop.rchain.comm.transport._
 import coop.rchain.comm.discovery._
 import coop.rchain.shared._, ThrowableOps._
 import coop.rchain.node.api._
+import coop.rchain.node.connect.Connect
 
 import scala.io.Source
 import scala.util.{Failure, Success, Try}
@@ -261,7 +262,7 @@ class NodeRuntime(conf: Conf)(implicit scheduler: Scheduler) {
   def handleCommunications: ProtocolMessage => Effect[CommunicationResponse] =
     pm =>
       NodeDiscovery[Effect].handleCommunications(pm) >>= {
-        case NotHandled => p2p.Network.dispatch[Effect](pm)
+        case NotHandled => Connect.dispatch[Effect](pm)
         case handled    => handled.pure[Effect]
     }
 
@@ -280,8 +281,8 @@ class NodeRuntime(conf: Conf)(implicit scheduler: Scheduler) {
               else
                 conf.run.bootstrap.toOption
                   .fold[Either[CommError, String]](Left(BootstrapNotProvided))(Right(_))
-                  .toEffect >>= (addr => p2p.Network.connectToBootstrap[Effect](addr)))
-      _ <- if (res.isRight) MonadOps.forever(p2p.Network.findAndConnect[Effect], 0)
+                  .toEffect >>= (addr => Connect.connectToBootstrap[Effect](addr)))
+      _ <- if (res.isRight) MonadOps.forever(Connect.findAndConnect[Effect], 0)
           else ().pure[Effect]
       _ <- casperEffect.close()
       _ <- exit0.toEffect
